+++ src/Autoversioning.php
@@ -1,52 +1,36 @@
 <?php
-/**
- * Craft3 Assets Autoversioning plugin for Craft CMS 3.x
- *
- * A Twig extension for CraftCMS (Craft3.x) that helps you cache-bust your assets
- *
- * @link      https://www.codemonauts.com
- * @copyright Copyright (c) 2018 Codemonauts
- */
 
 namespace codemonauts\autoversioning;
 
+use codemonauts\autoversioning\models\Settings;
 use codemonauts\autoversioning\twigextensions\AutoversioningTwigExtension;
 
 use Craft;
+use craft\base\Model;
 use craft\base\Plugin;
-use craft\services\Plugins;
-use craft\events\PluginEvent;
+use craft\helpers\UrlHelper;
 
-use yii\base\Event;
-
-/**
- * Class Craft3AssetsAutoversioning
- *
- * @author    Codemonauts
- * @package   Craft3AssetsAutoversioning
- * @since     0.1
- *
- */
 class Autoversioning extends Plugin
 {
-    // Static Properties
-    // =========================================================================
-
     /**
      * @var Autoversioning
      */
-    public static $plugin;
+    public static Autoversioning $plugin;
 
-    // Public Properties
-    // =========================================================================
+    /**
+     * @var \codemonauts\autoversioning\models\Settings|null
+     */
+    public static ?Settings $settings;
+
+    /**
+     * @inheritDoc
+     */
+    public bool $hasCpSettings = true;
 
     /**
      * @var string
      */
-    public $schemaVersion = '0.1';
-
-    // Public Methods
-    // =========================================================================
+    public string $schemaVersion = '0.1';
 
     /**
      * @inheritdoc
@@ -54,30 +38,47 @@ class Autoversioning extends Plugin
     public function init()
     {
         parent::init();
+
         self::$plugin = $this;
 
+        self::$settings = self::$plugin->getSettings();
+
+        // Register Twig extension
         Craft::$app->view->registerTwigExtension(new AutoversioningTwigExtension());
-
-        Event::on(
-            Plugins::class,
-            Plugins::EVENT_AFTER_INSTALL_PLUGIN,
-            function (PluginEvent $event) {
-                if ($event->plugin === $this) {
-                }
-            }
-        );
-
-        Craft::info(
-            Craft::t(
-                'craft3-assets-autoversioning',
-                '{name} plugin loaded',
-                ['name' => $this->name]
-            ),
-            __METHOD__
-        );
     }
 
-    // Protected Methods
-    // =========================================================================
+    /**
+     * @inheritDoc
+     */
+    public function afterInstall(): void
+    {
+        parent::afterInstall();
 
+        if (Craft::$app->getRequest()->getIsConsoleRequest()) {
+            return;
+        }
+
+        Craft::$app->getResponse()->redirect(
+            UrlHelper::cpUrl('settings/plugins/craft3-assets-autoversioning')
+        )->send();
+    }
+
+    /**
+     * @inheritDoc
+     */
+    protected function createSettingsModel(): ?Model
+    {
+        return new Settings();
+    }
+
+    /**
+     * @inheritDoc
+     */
+    protected function settingsHtml(): ?string
+    {
+        return Craft::$app->getView()->renderTemplate('craft3-assets-autoversioning/settings', [
+                'settings' => $this->getSettings(),
+            ]
+        );
+    }
 }