+++ CHANGELOG.md
@@ -1,22 +1,40 @@
 # Assets Autoversioning Changelog
 
-## Unreleased
+## 2.0.0 - 2022-05-04
+
+### Added
+
+- Craft CMS 4 compatibility.
+- Settings page.
+- Switch to activate/deactivate versioning. Can be set via environment variable.
+- The file with the build ID to use as version is now configurable.
+- Use query parameter instead of file suffix when adding the version to an asset.
+
 ### Changed
-- Fixed background color of icon
+
+- Requires Craft CMS >= 4.0
 
 ## 1.0.3 - 2019-04-06
+
 ### Changed
+
 - New icon
 - Updated documentation
 
 ## 1.0.2 - 2019-01-09
+
 ### Removed
+
 - Removed version number
 
 ## 1.0.1 - 2018-12-19
+
 ### Changed
+
 - Fix version number
 
 ## 1.0.0 - 2018-12-19
+
 ### Added
+
 - Initial release